Fix file extension and enable source generators for data editors
