+++ StrideLiveEditor.Avalonia/DataTypeEditors/QuaternionEditor.axaml
@@ -2,7 +2,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:local="clr-namespace:StrideLiveEditor.DataTypeEditors"
+             xmlns:local="using:StrideLiveEditor.Avalonia.DataTypeEditors"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:Class="StrideLiveEditor.Avalonia.DataTypeEditors.QuaternionEditor">
   <Grid>
