+++ StrideLiveEditor.Avalonia/DataTypeEditors/Vector3Editor.axaml
@@ -2,9 +2,9 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:local="clr-namespace:StrideLiveEditor.DataTypeEditors"
+             xmlns:local="using:StrideLiveEditor.Avalonia.DataTypeEditors"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="StrideLiveEditor.Avalonia.DataTypeEditors.Vector2Editor">
+             x:Class="StrideLiveEditor.Avalonia.DataTypeEditors.Vector3Editor">
   <Grid>
     <Grid.ColumnDefinitions>
       <ColumnDefinition Width="1*" />
@@ -17,11 +17,15 @@
         <ColumnDefinition Width="*" />
         <ColumnDefinition Width="10" />
         <ColumnDefinition Width="*" />
+        <ColumnDefinition Width="10" />
+        <ColumnDefinition Width="*" />
       </Grid.ColumnDefinitions>
       <TextBlock Text="X:" VerticalAlignment="Center" Grid.Column="0"/>
       <NumericUpDown x:Name="X" Grid.Column="1" ParsingNumberStyle="Float" FormatString="0.####" Increment="0.1" AllowSpin="True" Margin="5" />
       <TextBlock Text="Y:" VerticalAlignment="Center" Grid.Column="2"/>
       <NumericUpDown x:Name="Y" Grid.Column="3" ParsingNumberStyle="Float" FormatString="0.####" Increment="0.1" AllowSpin="True" Margin="5" />
+      <TextBlock Text="Z:" VerticalAlignment="Center" Grid.Column="4"/>
+      <NumericUpDown x:Name="Z" Grid.Column="5" ParsingNumberStyle="Float" FormatString="0.####" Increment="0.1" AllowSpin="True" Margin="5" />
     </Grid>
   </Grid>
 </local:BaseEditor>
